requirements.txt
blinker==1.7.0
cffi==1.16.0
click==8.1.7
cryptography==42.0.5
dnspython==2.4.2
Flask==3.0.0
Flask-JWT-Extended==4.5.3
itsdangerous==2.1.2
Jinja2==3.1.2
jwcrypto==1.5.6
MarkupSafe==2.1.3
mongoengine==0.27.0
pycparser==2.21
PyJWT==2.8.0
pymongo==4.6.0
python-dotenv==1.0.0
typing_extensions==4.10.0
Werkzeug==3.0.1
flask_cors==4.0.0
waitress==2.1.2
gevent==23.7.0
requests==2.31.0
colorlog==6.7.0
marshmallow
passlib